+++ src/main/java/portfolio/StudentManagement/StudentManagementApplication.java
@@ -6,6 +6,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import portfolio.StudentManagement.data.Student;
+import portfolio.StudentManagement.data.StudentCourse;
+import portfolio.StudentManagement.repository.StudentCourseRepository;
+import portfolio.StudentManagement.repository.StudentRepository;
 
 @SpringBootApplication
 @RestController
@@ -20,13 +24,5 @@ public class StudentManagementApplication {
 		SpringApplication.run(StudentManagementApplication.class, args);
 	}
 
-	@GetMapping("/studentList")
-	public List<Student> getStudentList() {
-		return studentRepository.selectAllStudentList();
-	}
 
-	@GetMapping("/studentCourseList")
-	public List<StudentCourse> getStudentCourseList() {
-		return studentCourseRepository.selectAllStudentCourseList();
-	}
 }
