+++ src/main/java/portfolio/StudentManagement/repository/StudentCourseRepository.java
@@ -1,8 +1,9 @@
-package portfolio.StudentManagement;
+package portfolio.StudentManagement.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import portfolio.StudentManagement.data.StudentCourse;
 
 @Mapper
 public interface StudentCourseRepository {
