+++ src/main/java/portfolio/StudentManagement/service/StudentService.java
@@ -1,0 +1,30 @@
+package portfolio.StudentManagement.service;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import portfolio.StudentManagement.data.Student;
+import portfolio.StudentManagement.data.StudentCourse;
+import portfolio.StudentManagement.repository.StudentCourseRepository;
+import portfolio.StudentManagement.repository.StudentRepository;
+
+@Service
+public class StudentService {
+  private StudentRepository studentRepository;
+  private StudentCourseRepository studentCourseRepository;
+
+  @Autowired
+  public StudentService(StudentRepository studentRepository,
+      StudentCourseRepository studentCourseRepository) {
+    this.studentRepository = studentRepository;
+    this.studentCourseRepository = studentCourseRepository;
+  }
+
+  public List<Student> selectAllStudentList() {
+    return studentRepository.selectAllStudentList();
+  }
+
+  public List<StudentCourse> gelectAllStudentCourseList() {
+    return studentCourseRepository.selectAllStudentCourseList();
+  }
+}
